requirements.txt
openai
python-dotenv
replicate
requests
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# Type stubs
google-api-python-client-stubs
google-auth-stubs
google-auth-oauthlib-stubs
types-requests

# Testing
pytest
